moved all queries.js files into lib directory. Edited routers to use new file locations. Currently working as intended
