--- db/seeds/05_widgets.sql
@@ -1,4 +1,0 @@
--- Widgets table seeds here (Example)
-INSERT INTO widgets (name, user_id) VALUES ('Sprockets', 1);
-INSERT INTO widgets (name, user_id) VALUES ('Chains', 2);
-INSERT INTO widgets (name, user_id) VALUES ('Bearings', 2);
