+++ lib/db.js
@@ -1,8 +1,13 @@
-let dbParams = {};
+// PG database client/connection setup
+const pg = require('pg');
+const Client = pg.Client;
+
+let config = {};
+// If connected to Heroku Database:
 if (process.env.DATABASE_URL) {
-  dbParams.connectionString = process.env.DATABASE_URL;
+  config.connectionString = process.env.DATABASE_URL;
 } else {
-  dbParams = {
+  config = {
     host: process.env.DB_HOST,
     port: process.env.DB_PORT,
     user: process.env.DB_USER,
@@ -11,4 +16,11 @@ if (process.env.DATABASE_URL) {
   };
 }
 
-module.exports = dbParams;
+const client = new Client(config);
+
+client.connect(() => {
+  console.log('connected to database');
+});
+
+module.exports = client;
+
