+++ lib/queriesFavorites.js
@@ -1,4 +1,4 @@
-const db = require('./testDB');
+const db = require('./db');
 
 const addFavorite = (user_id, map_id) => {
