+++ lib/queriesMaps.js
@@ -1,4 +1,4 @@
-const db = require('./testDB');
+const db = require('./db');
 
 const getMaps = () => {
   return db.query('SELECT * FROM maps LIMIT 10;')
