+++ lib/queriesUsers.js
@@ -1,4 +1,4 @@
-const db = require('./testDB');
+const db = require('./db');
 
 const getUserById = (id) => {
   return db.query('SELECT * FROM users WHERE name = $1', [id])
