+++ routes/maps-router.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router  = express.Router();
-const { getMaps, getMapById } = require('../db/testQueriesMaps');
+const { getMaps, getMapById } = require('../lib/queriesMaps');
 
 //  /GET/maps
 router.get('/', (req, res) => {
